refactor: separate sruct functions from admin page
