+++ client/src/pages/AdminPage.js
@@ -4,23 +4,11 @@ import TabPanel from "../components/Tabs/TabPanel";
 import VerticalTabs from "../components/Tabs/VerticalTabs";
 import CountryAddForm from "../components/Forms/CountryAddForm";
 import CountryUpdateForm from "../components/Forms/CountryUpdateForm";
+import {destructCountry, structCountries} from "../helpers/struct"
 import classesCss from './styles/AdminPage.module.scss'
-import Map from "../components/Map/Map";
 
 
-const langShortDummy = {
-    lang: "",
-    countryName: "",
-    shortText: "",
-    capitalName: "",
-}
-const langFullDummy = {
-    lang: "",
-    countryCode: "",
-    countryName: "",
-    description: "",
-    capitalName: "",
-}
+
 
 const AdminPage = () => {
     const {request, loading} = useHttp()
@@ -47,73 +35,6 @@ const AdminPage = () => {
         }
     }
 
-    const destructCountry= (countryDataFromClient) => {
-        const countryStructure = {
-            ...countryDataFromClient.data,
-            langData: []
-        }
-        const langCountryStructure = []
-        for(let lang in countryDataFromClient.data.langData){
-            if(countryDataFromClient.data.langData.hasOwnProperty(lang)){
-                const shortLangData = {lang}
-                const fullLangData = {lang, countryCode: countryDataFromClient.countryCode}
-                for(let shortDummyKey in langShortDummy){
-                    if(countryDataFromClient.data.langData[lang][shortDummyKey] !== undefined){
-                        shortLangData[shortDummyKey] = countryDataFromClient.data.langData[lang][shortDummyKey]
-                    }
-
-                }
-                for(let fullDummyKey in langFullDummy){
-                    if(countryDataFromClient.data.langData[lang][fullDummyKey] !== undefined){
-                        fullLangData[fullDummyKey] = countryDataFromClient.data.langData[lang][fullDummyKey]
-                    }
-                }
-                countryStructure.langData.push(shortLangData);
-                langCountryStructure.push(fullLangData)
-            }
-        }
-
-        return {
-            countryData: countryStructure,
-            langCountryData: langCountryStructure
-        }
-    }
-
-    const structCountries = (countriesDataFromServer) => {
-        const structure = {}
-        for(let i = 0; i < countriesDataFromServer.countries.length; i++){
-
-            const countryCode = countriesDataFromServer.countries[i].countryCode
-
-            structure[countryCode] = countriesDataFromServer.countries[i]
-            // delete structure[countryCode].__id
-            // delete structure[countryCode].__v
-
-            const countryLangData = {};
-            for (let i = 0; i < structure[countryCode].langData.length; i++) {
-                countryLangData[structure[countryCode].langData[i].lang] = structure[countryCode].langData[i]
-            }
-            if(countriesDataFromServer.langCountries){
-                const currentCountryLangs = countriesDataFromServer.langCountries.filter(county => {
-                    return county.countryCode === countryCode
-                })
-                for (let i = 0; i < currentCountryLangs.length; i++) {
-                    countryLangData[currentCountryLangs[i].lang] = {
-                        ...countryLangData[currentCountryLangs[i].lang],
-                        ...currentCountryLangs[i]
-                    }
-                    delete countryLangData[currentCountryLangs[i].lang].lang
-                    delete countryLangData[currentCountryLangs[i].lang].countryCode
-                }
-
-            }
-            structure[countryCode].langData = countryLangData
-
-
-        }
-        return structure
-    }
-
 
     useEffect(() => {
         getCountryFromBase({})
@@ -124,6 +45,8 @@ const AdminPage = () => {
             setCountriesData(structCountries(countryResponse))
         }
     }, [countryResponse])
+
+    console.log(countryResponse)
 
     return (
         <>
